+++ src/main/java/com/example/team1_be/domain/Schedule/Service/ScheduleService.java
@@ -38,8 +38,8 @@ import com.example.team1_be.domain.User.UserService;
 import com.example.team1_be.domain.Week.Service.WeekService;
 import com.example.team1_be.domain.Week.Week;
 import com.example.team1_be.domain.Week.WeekRecruitmentStatus;
+import com.example.team1_be.domain.Worktime.Service.WorktimeService;
 import com.example.team1_be.domain.Worktime.Worktime;
-import com.example.team1_be.domain.Worktime.WorktimeService;
 import com.example.team1_be.utils.errors.exception.CustomException;
 import com.example.team1_be.utils.errors.exception.NotFoundException;
